fixed new user issue

+++ backend/app.js
@@ -53,33 +53,33 @@ app.get('/users', async (req, res) => {
 });
 
 // Endpoint to handle form submission
+// Endpoint to handle form submission or user profile update
 app.post('/updateUser', upload.single('photo'), async (req, res) => {
   try {
      // Convert the uploaded file to Base64
-     const photoBase64 = fs.readFileSync(req.file.path, 'base64');
+     const photoBase64 = req.file ? fs.readFileSync(req.file.path, 'base64') : null;
  
-     // Check if a user document exists with the given UID
-     const existingUser = await User.findOne({ uid: req.body.uid });
+     // Check if a user document exists with the given email
+     const existingUser = await User.findOne({ email: req.body.email });
  
      let user;
      if (existingUser) {
        // If the user exists, update the document
        user = await User.findOneAndUpdate(
-         { uid: req.body.uid },
+         { email: req.body.email },
          {
            username: req.body.username,
-           photo: photoBase64,
-           // Add logic here to update the providers array if necessary
+           photo: photoBase64 || existingUser.photo, // Use the existing photo if no new photo is uploaded
+           updated_at: new Date(),
          },
          { new: true } // Return the updated document
        );
      } else {
        // If the user does not exist, create a new document
        user = new User({
-         uid: req.body.uid,
+         email: req.body.email,
          username: req.body.username,
          photo: photoBase64,
-         // Initialize the providers array with the current authMethod
          providers: [{ provider: req.body.authMethod }],
          created_at: new Date(),
          updated_at: new Date(),
@@ -88,13 +88,15 @@ app.post('/updateUser', upload.single('photo'), async (req, res) => {
      }
  
      // Delete the file after it's converted to Base64 and saved to the database
-     fs.unlink(req.file.path, (err) => {
-       if (err) {
-         console.error('Error deleting file:', err);
-       } else {
-         console.log('File deleted successfully');
-       }
-     });
+     if (req.file) {
+       fs.unlink(req.file.path, (err) => {
+         if (err) {
+           console.error('Error deleting file:', err);
+         } else {
+           console.log('File deleted successfully');
+         }
+       });
+     }
  
      res.status(200).json({ message: 'Data saved successfully', user });
   } catch (error) {
@@ -102,6 +104,7 @@ app.post('/updateUser', upload.single('photo'), async (req, res) => {
      res.status(500).json({ message: 'Server error' });
   }
  });
+
